fully functional double click on local vid to swap between grid and popout
